src/rag/requirements.txt
# Essentials only for basic RAG functionality
python-dotenv>=1.0.0
requests>=2.25.0
flask>=2.0.0
flask-cors>=3.0.0

# Text processing
pypdf2>=3.0.0
python-docx>=0.8.11

# Simple vector search (alternative to complex dependencies)
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Optional: Simple OpenAI client (if needed)
# openai>=1.0.0